Moved GetOrNoneManger to model utils and made changes to get_invitation_obj as per suggestions from kaapstorm
