+++ corehq/apps/consumer_user/models.py
@@ -1,17 +1,7 @@
 from django.contrib.auth.models import User
 from django.db import models
 
-
-class GetOrNoneManager(models.Manager):
-    """
-    Adds get_or_none method to objects
-    """
-
-    def get_or_none(self, **kwargs):
-        try:
-            return self.get(**kwargs)
-        except self.model.DoesNotExist:
-            return None
+from corehq.util.models import GetOrNoneManager
 
 
 class ConsumerUser(models.Model):
